A LOT of refactorization and removal of unneccesary junk.
