+++ src/main/java/com/stuypulse/robot/Main.java
@@ -2,7 +2,6 @@
 /* Copyright (c) 2024 StuyPulse Robotics. All rights reserved.*/
 /* This work is licensed under the terms of the MIT license.  */
 /**************************************************************/
-
 package com.stuypulse.robot;
 
 import edu.wpi.first.wpilibj.RobotBase;
