+++ src/main/java/com/stuypulse/robot/RobotContainer.java
@@ -1,25 +1,20 @@
 package com.stuypulse.robot;
 
-import com.stuypulse.robot.commands.auton.Combinations.PushBackwardsL1Auton;
-
-import com.stuypulse.robot.commands.auton.CoralOnly.DoubleCoralAuton;
-import com.stuypulse.robot.commands.auton.CoralOnly.SingleCoralAuton;
-
+import com.stuypulse.robot.commands.auton.Combinations.PushBackwardsCoralAuton;
+import com.stuypulse.robot.commands.auton.OnlyCoral.DoubleCoralAuton;
+import com.stuypulse.robot.commands.auton.OnlyCoral.SingleCoralAuton;
 import com.stuypulse.robot.commands.auton.Misc.DoNothingAuton;
 import com.stuypulse.robot.commands.auton.Misc.MobilityAuton;
-
-import com.stuypulse.robot.commands.auton.PushOnly.PushBackwardsAuton;
-import com.stuypulse.robot.commands.auton.PushOnly.PushForwardsAuton;
-
+import com.stuypulse.robot.commands.auton.OnlyPush.PushBackwardsAuton;
+import com.stuypulse.robot.commands.auton.OnlyPush.PushForwardsAuton;
 import com.stuypulse.robot.commands.climb.ClimbToClimb;
 import com.stuypulse.robot.commands.climb.ClimbToStow;
-
+import com.stuypulse.robot.commands.drive.DriveDefault;
 import com.stuypulse.robot.commands.leds.LEDApplyPattern;
 import com.stuypulse.robot.commands.leds.LEDDeafultCommand;
 import com.stuypulse.robot.commands.pivot.PivotLower;
 import com.stuypulse.robot.commands.pivot.PivotRaise;
 import com.stuypulse.robot.commands.pivot.PivotResetAngle;
-import com.stuypulse.robot.commands.pivot.PivotStop;
 import com.stuypulse.robot.commands.pivot.PivotToAlgaeIntake;
 import com.stuypulse.robot.commands.pivot.PivotToAlgaeStow;
 import com.stuypulse.robot.commands.pivot.PivotToCoralStow;
@@ -29,10 +24,10 @@ import com.stuypulse.robot.commands.pivot.roller.PivotCoralOuttake;
 import com.stuypulse.robot.commands.pivot.roller.PivotHoldCoral;
 import com.stuypulse.robot.commands.pivot.roller.PivotRollerStop;
 
-import com.stuypulse.robot.constants.Ports;
 import com.stuypulse.robot.subsystems.drivetrain.Drivetrain;
 import com.stuypulse.robot.subsystems.leds.LEDController;
 import com.stuypulse.robot.subsystems.pivot.Pivot;
+import com.stuypulse.robot.constants.Ports;
 
 import com.stuypulse.stuylib.input.Gamepad;
 import com.stuypulse.stuylib.input.gamepads.AutoGamepad;
@@ -43,7 +38,6 @@ import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.sysid.SysIdRoutine;
 
 public class RobotContainer {
-
     // Gamepads
     public final Gamepad driver = new AutoGamepad(Ports.Gamepad.DRIVER);
     public final Gamepad operator = new AutoGamepad(Ports.Gamepad.OPERATOR);
@@ -58,9 +52,7 @@ public class RobotContainer {
     // Autons
     private static SendableChooser<Command> autonChooser = new SendableChooser<>();
 
-
     // Robot container
-
     public RobotContainer() {
         configureDefaultCommands();
         configureButtonBindings();
@@ -78,46 +70,47 @@ public class RobotContainer {
         driveSubsystem.setDefaultCommand(new DriveDefault(driver, true));
     }
 
-    /***************/
-    /*** BUTTONS ***/
-    /***************/
+    /***********************/
+    /*** BUTTON BINDINGS ***/
+    /***********************/
 
     private void configureButtonBindings() {
-            driver.getTopButton()
-                .whileTrue(new PivotRollerStop())
-                .whileTrue(new PivotCoralOuttake())
-                .onFalse(new PivotHoldCoral());
-            driver.getLeftButton()
-                .whileTrue(new ClimbToClimb());
-            driver.getRightButton()
-                .whileTrue(new ClimbToStow());
-            // driver.getBottomButton()
-            //     .whileTrue(new VisionAlignToReef())
+        //BUMPERS
+        driver.getTopButton()
+            .whileTrue(new PivotRollerStop())
+            .whileTrue(new PivotCoralOuttake())
+            .onFalse(new PivotHoldCoral());
+        driver.getLeftButton()
+            .whileTrue(new ClimbToClimb());
+        driver.getRightButton()
+            .whileTrue(new ClimbToStow());
+        // driver.getBottomButton()
+        //     .whileTrue(new VisionAlignToReef())
 
-            driver.getRightTriggerButton()
-                .whileTrue(new PivotRaise())
-                .onFalse(new PivotStop());
-            driver.getLeftTriggerButton()
-                .whileTrue(new PivotLower())
-                .onFalse(new PivotStop());
+        //TRIGGERS
+        driver.getRightTriggerButton()
+            .whileTrue(new PivotRaise())
+            .onFalse(new PivotRollerStop());
+        driver.getLeftTriggerButton()
+            .whileTrue(new PivotLower())
+            .onFalse(new PivotRollerStop());
 
-            driver.getRightBumper() 
-                .whileTrue(new PivotAlgaeOuttake())
-                .onFalse(new PivotHoldCoral());
-            driver.getLeftBumper()
-                .whileTrue(new PivotAlgaeIntake())
-                .onFalse(new PivotRollerStop());
+        //BUMPERS
+        driver.getRightBumper() 
+            .whileTrue(new PivotAlgaeOuttake())
+            .onFalse(new PivotHoldCoral());
+        driver.getLeftBumper()
+            .whileTrue(new PivotAlgaeIntake())
+            .onFalse(new PivotRollerStop());
 
-            driver.getDPadUp()
-                .onTrue(new PivotToCoralStow());
-    
-            driver.getDPadRight()
-                .onTrue(new PivotToAlgaeStow());
-
-            driver.getDPadDown()
-                .onTrue(new PivotToAlgaeIntake());
-
-        }
+        //DPAD
+        driver.getDPadUp()
+            .onTrue(new PivotToCoralStow());
+        driver.getDPadRight()
+            .onTrue(new PivotToAlgaeStow());
+        driver.getDPadDown()
+            .onTrue(new PivotToAlgaeIntake());
+    }
         /*else if(Settings.DriveMode.GAMEPAD.toString() == "JOYSTICK") {     
             joystick.getTriggerTriggered()
                 .whileTrue(new PivotCoralOut())
@@ -166,18 +159,20 @@ public class RobotContainer {
         autonChooser.addOption("Push Only - Forwards", new PushForwardsAuton());
         autonChooser.addOption("Push Only - Backwards", new PushBackwardsAuton());
  
-        autonChooser.addOption("Coral - Single - With Push", new PushBackwardsL1Auton());
+        autonChooser.addOption("Combination - Coral w/ Push", new PushBackwardsCoralAuton());
 
         SmartDashboard.putData("Autonomous", autonChooser);
     }
+
     public void configureSysId() {
         SysIdRoutine pivotSysIdRoutine = pivot.getSysIdRoutine();
+        SysIdRoutine driveSysIdRoutine = driveSubsystem.getSysIdRoutine();
+        
         autonChooser.addOption("z-SysID - Pivot - Dynamic Forwards (Up)", pivotSysIdRoutine.dynamic(SysIdRoutine.Direction.kForward));
         autonChooser.addOption("z-SysID - Pivot - Dynamic Backwards (Down)", pivotSysIdRoutine.dynamic(SysIdRoutine.Direction.kReverse));
         autonChooser.addOption("z-SysID - Pivot - Quasistatic Forwards (Up)", pivotSysIdRoutine.quasistatic(SysIdRoutine.Direction.kForward));
         autonChooser.addOption("z-SysID - Pivot - Quasistatic Backwards (Down)", pivotSysIdRoutine.quasistatic(SysIdRoutine.Direction.kReverse));
 
-        SysIdRoutine driveSysIdRoutine = driveSubsystem.getSysIdRoutine();
         autonChooser.addOption("z-SysID - Drive - Dynamic Forward", driveSysIdRoutine.dynamic(SysIdRoutine.Direction.kForward));
         autonChooser.addOption("z-SysID - Drive - Dynamic Backwards", driveSysIdRoutine.dynamic(SysIdRoutine.Direction.kReverse));
         autonChooser.addOption("z-SysID - Drive - Quasistatic Forward", driveSysIdRoutine.quasistatic(SysIdRoutine.Direction.kForward));
