+++ src/main/java/com/stuypulse/robot/commands/DoNothingCommand.java
@@ -2,18 +2,16 @@
 /* Copyright (c) 2024 StuyPulse Robotics. All rights reserved.*/
 /* This work is licensed under the terms of the MIT license.  */
 /**************************************************************/
-
 package com.stuypulse.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 
-/*-
+/**
  * This command does a whole load of nothing...
  *
  * @author Ivan Chen
  */
 public class DoNothingCommand extends InstantCommand {
-
     public DoNothingCommand() {
         // Do loads of nothing
     }
