+++ src/main/java/com/stuypulse/robot/commands/auton/Misc/MobilityAuton.java
@@ -2,7 +2,6 @@
 /* Copyright (c) 2024 StuyPulse Robotics. All rights reserved.*/
 /* This work is licensed under the terms of the MIT license.  */
 /**************************************************************/
-
 package com.stuypulse.robot.commands.auton.Misc;
 
 import edu.wpi.first.wpilibj2.command.WaitUntilCommand;
@@ -13,7 +12,6 @@ import com.stuypulse.robot.commands.leds.LEDApplyPattern;
 import com.stuypulse.robot.constants.Settings.LEDPatterns;
 
 public class MobilityAuton extends SequentialCommandGroup {
-
     public MobilityAuton() {
         addCommands(
             new LEDApplyPattern(LEDPatterns.MOBILITY_AUTON),
