+++ src/main/java/com/stuypulse/robot/commands/climb/ClimbToClimb.java
@@ -3,7 +3,7 @@ package com.stuypulse.robot.commands.climb;
 import com.stuypulse.robot.subsystems.climber.Climb.ClimbState;
 
 public class ClimbToClimb extends ClimbToState{
-    public ClimbToClimb(){
+    public ClimbToClimb() {
         super(ClimbState.CLIMBING);
     }
 }
