+++ src/main/java/com/stuypulse/robot/commands/climb/ClimbToStow.java
@@ -2,7 +2,6 @@ package com.stuypulse.robot.commands.climb;
 
 import com.stuypulse.robot.subsystems.climber.Climb.ClimbState;
 
-
 public class ClimbToStow extends ClimbToState {
     public ClimbToStow() {
         super(ClimbState.STOW);
