+++ src/main/java/com/stuypulse/robot/commands/drive/DrivePIDToPose.java
@@ -18,20 +18,11 @@ public class DrivePIDToPose extends Command {
 
         addRequirements(Drivetrain.getInstance());
     }
-    
-    @Override
-    public void initialize() {}
 
     @Override
     public void execute() {
-        double currentSpeed = (Drivetrain.getInstance().getRightVelocity() + Drivetrain.getInstance().getLeftVelocity())/2;
-        double speedDifference = currentSpeed-xSpeed;
-
-        Drivetrain.getInstance().driveTank(currentSpeed, currentSpeed, squared);
+        //TODO: code driving to a Pose2d
     }
-
-    @Override
-    public void end(boolean interrupted) {}
 
     @Override
     public boolean isFinished() {
