+++ src/main/java/com/stuypulse/robot/commands/pivot/PivotLower.java
@@ -3,9 +3,7 @@ package com.stuypulse.robot.commands.pivot;
 import com.stuypulse.robot.constants.Settings;
 
 public class PivotLower extends PivotToDirection {
-
     public PivotLower() {
-
         super(Settings.Pivot.PIVOT_LOWER_SPEED.doubleValue());
     }
 }
