+++ src/main/java/com/stuypulse/robot/commands/pivot/PivotToDirection.java
@@ -8,9 +8,10 @@ public class PivotToDirection extends Command {
     private Pivot pivot; 
     private double pivotMotorSpeed;
   
-  public PivotToDirection(double pivotMotorSpeed) {
+    public PivotToDirection(double pivotMotorSpeed) {
         pivot = Pivot.getInstance();
         this.pivotMotorSpeed = pivotMotorSpeed;
+        
         addRequirements(pivot); 
     }
