+++ src/main/java/com/stuypulse/robot/commands/pivot/PivotToState.java
@@ -6,8 +6,18 @@ import com.stuypulse.robot.subsystems.pivot.Pivot.PivotState;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 
 public class PivotToState extends InstantCommand {
+    private Pivot pivot;
+    private PivotState pivotState;
+
     public PivotToState(PivotState pivotState) {
-        super(() -> Pivot.getInstance().setPivotState(pivotState));
-        addRequirements(Pivot.getInstance());
+        pivot = Pivot.getInstance();
+        this.pivotState = pivotState;
+
+        addRequirements(pivot);
+    }
+
+    @Override
+    public void initialize() {
+        pivot.setPivotState(pivotState);
     }
 }
