+++ src/main/java/com/stuypulse/robot/commands/pivot/roller/PivotRollerToDirection.java
@@ -7,14 +7,15 @@ import edu.wpi.first.wpilibj2.command.Command;
 public class PivotRollerToDirection extends Command{
     private Pivot pivot;
     private double pivotRollerSpeed;
+
     public PivotRollerToDirection(double pivotRollerSpeed) {
         pivot = Pivot.getInstance();
         this.pivotRollerSpeed = pivotRollerSpeed;
+
         addRequirements(pivot);
     }
 
     public void initialize() {
         pivot.setRollerMotor(pivotRollerSpeed);
     }
-
 }
