+++ src/main/java/com/stuypulse/robot/constants/Constants.java
@@ -2,8 +2,8 @@ package com.stuypulse.robot.constants;
 
 import edu.wpi.first.math.geometry.Rotation2d;
 
+/** Miscellaneous constants */
 public interface Constants {
-
     public interface Climb {
         Rotation2d MIN_ANGLE = Rotation2d.kZero;
         Rotation2d MAX_ANGLE = Rotation2d.fromDegrees(0.0);
@@ -12,6 +12,6 @@ public interface Constants {
     }
 
     public interface Drivetrain {
-        public static final double WHEEL_CIRCUMFERENCE = 0.479;
+        public static final double WHEEL_CIRCUMFERENCE = 0.479; //TODO: RECHECK WHEEL CIRCUMFERENCE, SHOULD BE IN INCHES
     }
 }
