+++ src/main/java/com/stuypulse/robot/constants/Gains.java
@@ -1,7 +1,9 @@
 package com.stuypulse.robot.constants;
 
+/** The class that contains controller values for all subsystems that require the use of 
+ *  one or multiple controllers such as Feed Forward (FF) and Proportional-Integral-Derivative (PID)
+ */
 public class Gains {
-    
     public interface Climb {
         public interface FF {
             double kS = 0.0;
@@ -18,10 +20,7 @@ public class Gains {
     }
 
     public interface Pivot {
-         public interface FF {
-            double kS = 0.0;
-            double kV = 0.0;
-            double kA = 0.0;
+         public interface FF { //NO kA, kV, kS. PID + kG only.
             double kG = 0.21;
         }
         public interface PID {
@@ -46,5 +45,3 @@ public class Gains {
     }
 }
     
-        
-
