+++ src/main/java/com/stuypulse/robot/subsystems/drivetrain/Drivetrain.java
@@ -1,18 +1,12 @@
-// Lucas O, Sebastian, Daniel, Yunus
-
 package com.stuypulse.robot.subsystems.drivetrain;
 
 import com.stuypulse.stuylib.control.Controller;
 import com.stuypulse.stuylib.control.angle.AngleController;
-import com.stuypulse.stuylib.control.feedback.PIDController;
-import com.stuypulse.stuylib.control.feedforward.MotorFeedforward;
 
-import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import edu.wpi.first.wpilibj2.command.sysid.SysIdRoutine;
 
 public abstract class Drivetrain extends SubsystemBase {
-
     private static final Drivetrain instance;
 
     public Controller controllerPosition;
