+++ src/main/java/com/stuypulse/robot/subsystems/leds/LEDController.java
@@ -1,9 +1,8 @@
 package com.stuypulse.robot.subsystems.leds;
 
-import edu.wpi.first.wpilibj2.command.SubsystemBase;
-
 import com.stuypulse.robot.constants.Ports;
 
+import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import edu.wpi.first.wpilibj.AddressableLED;
 import edu.wpi.first.wpilibj.AddressableLEDBuffer;
 import edu.wpi.first.wpilibj.LEDPattern;
